Move namespace unshare to strat_engine
